feat: Query detection in map statements, part 1
GraphQL needs to see where the root of a query is, so it can apply the correct selection.
It's not necessarily at the "root property" of the output struct.
When it _is_ at the root property, we want RelationshipId(def@0:0) (unit!) to represent this.
So change all the tests involving synthetic property_id to avoid using the ONTOL domain, to better reflect the real world.