fix: separate more things the compiler does
The compiler does checks by looping over the DefIds in domains.
This changes the architecture a bit so that there are overall more steps to perform,
there is less happening in each step, but each step involves looping over the definitions.

This leads to more deterministic behaviour, less chicken-and-egg problems.